Move query federation into quasar.evaluate package.
