+++ core/src/test/scala/quasar/evaluate/FederatingQueryEvaluatorSpec.scala
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package quasar
+package quasar.evaluate
 
 import slamdata.Predef._
+import quasar.{Qspec, TreeMatchers}
 import quasar.api._, ResourceError._
 import quasar.contrib.pathy.{ADir, AFile}
 import quasar.fp.{constEqual, coproductEqual}
